requirements.txt
django==4.2
djangorestframework
Pillow

django-cors-headers
rest_framework_simplejwt
python-dateutil